+++ tbx/cyanobacteria-tools/CellData.m
@@ -1,0 +1,14 @@
+classdef
+    
+    properties
+        
+        Channels
+        Parameters
+        
+    end
+    
+    methods
+        
+    end
+    
+end