+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 
 ## Version 0.0.1 (Release date: *TBA*)
 
+*2017-10-03*
+* Started writing TrackDataArray and TrackData classes
+
+*2017-10-02*
+* Started writing TrackLinker class
+
 *2017-09-30*
 * Started the project to consolidate the Cameron lab code. Most of the project have similar requirements, but are now spread out over multiple repositories. It would be a lot better to consolidate all the code so we only have to maintain a single codebase.